refactor(ui): Extract dropdown menu components into separate file
Move dropdown menu related components from properties-panel.tsx to a new
dedicated file dropdown-menu.tsx for better code organization and reusability.
The new file contains the following components:

- DropdownMenu
- DropdownMenuTrigger
- DropdownMenuContent
- DropdownMenuLabel
- DropdownMenuSeparator
- DropdownMenuRadioItem
- DropdownMenuCheckboxItem

No functional changes or breaking changes introduced.
